update/WH-2 - Update controller layer

+++ src/main/java/com/github/blog/controller/CommentController.java
@@ -1,12 +1,16 @@
 package com.github.blog.controller;
 
 import com.github.blog.controller.dto.common.CommentDto;
-import com.github.blog.controller.dto.request.CommentDtoFilter;
 import com.github.blog.controller.dto.request.CommentRequest;
 import com.github.blog.controller.dto.request.PageableRequest;
+import com.github.blog.controller.dto.request.filter.CommentDtoFilter;
 import com.github.blog.controller.dto.response.Page;
 import com.github.blog.service.CommentService;
+import com.github.blog.service.exception.UserErrorResult;
+import com.github.blog.service.exception.impl.UserException;
+import com.github.blog.service.security.impl.UserDetailsImpl;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -26,7 +30,10 @@ public class CommentController {
     private final CommentService commentService;
 
     @PostMapping
-    public CommentDto create(@RequestBody CommentRequest request) {
+    public CommentDto create(@RequestBody CommentRequest request, @AuthenticationPrincipal UserDetailsImpl userDetails) {
+        if (!userDetails.getId().equals(request.getUserId())) {
+            throw new UserException(UserErrorResult.UNAUTHORIZED_CREATE_ATTEMPT);
+        }
         return commentService.create(request);
     }
 
@@ -41,12 +48,18 @@ public class CommentController {
     }
 
     @PutMapping("{id}")
-    public CommentDto update(@PathVariable("id") Long id, @RequestBody CommentRequest request) {
+    public CommentDto update(@PathVariable("id") Long id, @RequestBody CommentRequest request, @AuthenticationPrincipal UserDetailsImpl userDetails) {
+        if (!userDetails.isAdmin() && !userDetails.getId().equals(id)) {
+            throw new UserException(UserErrorResult.UNAUTHORIZED_UPDATE_ATTEMPT);
+        }
         return commentService.update(id, request);
     }
 
     @DeleteMapping("{id}")
-    public CommentDto delete(@PathVariable("id") Long id) {
+    public CommentDto delete(@PathVariable("id") Long id, @AuthenticationPrincipal UserDetailsImpl userDetails) {
+        if (!userDetails.isAdmin() && !userDetails.getId().equals(id)) {
+            throw new UserException(UserErrorResult.UNAUTHORIZED_DELETION_ATTEMPT);
+        }
         return commentService.delete(id);
     }
 }
