+++ src/main/java/com/github/blog/controller/PostController.java
@@ -2,11 +2,15 @@ package com.github.blog.controller;
 
 import com.github.blog.controller.dto.common.PostDto;
 import com.github.blog.controller.dto.request.PageableRequest;
-import com.github.blog.controller.dto.request.PostDtoFilter;
 import com.github.blog.controller.dto.request.PostRequest;
+import com.github.blog.controller.dto.request.filter.PostDtoFilter;
 import com.github.blog.controller.dto.response.Page;
 import com.github.blog.service.PostService;
+import com.github.blog.service.exception.UserErrorResult;
+import com.github.blog.service.exception.impl.UserException;
+import com.github.blog.service.security.impl.UserDetailsImpl;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -26,7 +30,11 @@ public class PostController {
     private final PostService postService;
 
     @PostMapping
-    public PostDto create(@RequestBody PostRequest request) {
+    //@PreAuthorize("authentication.principal.id == request.userId")
+    public PostDto create(@RequestBody PostRequest request, @AuthenticationPrincipal UserDetailsImpl userDetails) {
+        if (!userDetails.getId().equals(request.getUserId())) {
+            throw new UserException(UserErrorResult.UNAUTHORIZED_CREATE_ATTEMPT);
+        }
         return postService.create(request);
     }
 
@@ -41,12 +49,18 @@ public class PostController {
     }
 
     @PutMapping("{id}")
-    public PostDto update(@PathVariable("id") Long id, @RequestBody PostRequest request) {
+    public PostDto update(@PathVariable("id") Long id, @RequestBody PostRequest request, @AuthenticationPrincipal UserDetailsImpl userDetails) {
+        if (!userDetails.isAdmin() && !userDetails.getId().equals(id)) {
+            throw new UserException(UserErrorResult.UNAUTHORIZED_UPDATE_ATTEMPT);
+        }
         return postService.update(id, request);
     }
 
     @DeleteMapping("{id}")
-    public PostDto delete(@PathVariable("id") Long id) {
+    public PostDto delete(@PathVariable("id") Long id, @AuthenticationPrincipal UserDetailsImpl userDetails) {
+        if (!userDetails.isAdmin() && !userDetails.getId().equals(id)) {
+            throw new UserException(UserErrorResult.UNAUTHORIZED_DELETION_ATTEMPT);
+        }
         return postService.delete(id);
     }
 }
