+++ src/main/java/com/github/blog/controller/PostReactionController.java
@@ -1,8 +1,13 @@
 package com.github.blog.controller;
 
 import com.github.blog.controller.dto.common.PostReactionDto;
+import com.github.blog.controller.dto.request.PostReactionRequest;
 import com.github.blog.service.PostReactionService;
+import com.github.blog.service.exception.UserErrorResult;
+import com.github.blog.service.exception.impl.UserException;
+import com.github.blog.service.security.impl.UserDetailsImpl;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -24,8 +29,11 @@ public class PostReactionController {
     private final PostReactionService postReactionService;
 
     @PostMapping
-    public PostReactionDto create(@RequestBody PostReactionDto postReactionDto) {
-        return postReactionService.create(postReactionDto);
+    public PostReactionDto create(@RequestBody PostReactionRequest request, @AuthenticationPrincipal UserDetailsImpl userDetails) {
+        if (!userDetails.getId().equals(request.getUserId())) {
+            throw new UserException(UserErrorResult.UNAUTHORIZED_CREATE_ATTEMPT);
+        }
+        return postReactionService.create(request);
     }
 
     @GetMapping("{id}")
@@ -39,12 +47,18 @@ public class PostReactionController {
     }
 
     @PutMapping("{id}")
-    public PostReactionDto update(@PathVariable("id") Long id, @RequestBody PostReactionDto postReactionDto) {
-        return postReactionService.update(id, postReactionDto);
+    public PostReactionDto update(@PathVariable("id") Long id, @RequestBody PostReactionRequest request, @AuthenticationPrincipal UserDetailsImpl userDetails) {
+        if (!userDetails.isAdmin() && !userDetails.getId().equals(id)) {
+            throw new UserException(UserErrorResult.UNAUTHORIZED_UPDATE_ATTEMPT);
+        }
+        return postReactionService.update(id, request);
     }
 
     @DeleteMapping("{id}")
-    public PostReactionDto delete(@PathVariable("id") Long id) {
+    public PostReactionDto delete(@PathVariable("id") Long id, @AuthenticationPrincipal UserDetailsImpl userDetails) {
+        if (!userDetails.isAdmin() && !userDetails.getId().equals(id)) {
+            throw new UserException(UserErrorResult.UNAUTHORIZED_DELETION_ATTEMPT);
+        }
         return postReactionService.delete(id);
     }
 }
