+++ src/main/java/com/github/blog/controller/UserInfoController.java
@@ -2,7 +2,12 @@ package com.github.blog.controller;
 
 import com.github.blog.controller.dto.common.UserInfoDto;
 import com.github.blog.service.UserInfoService;
+import com.github.blog.service.exception.UserErrorResult;
+import com.github.blog.service.exception.impl.UserException;
+import com.github.blog.service.security.impl.UserDetailsImpl;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -18,33 +23,46 @@ import java.util.List;
  * @author Raman Haurylau
  */
 @RestController
-@RequestMapping("user-details")
+@RequestMapping("user-info")
 @RequiredArgsConstructor
 public class UserInfoController {
     private final UserInfoService userInfoService;
 
     @PostMapping
-    public UserInfoDto create(@RequestBody UserInfoDto userDetails) {
-        return userInfoService.create(userDetails);
+    public UserInfoDto create(@RequestBody UserInfoDto request, @AuthenticationPrincipal UserDetailsImpl userDetails) {
+        if (!userDetails.getId().equals(request.getId())) {
+            throw new UserException(UserErrorResult.UNAUTHORIZED_CREATE_ATTEMPT);
+        }
+        return userInfoService.create(request);
     }
 
     @GetMapping("{id}")
-    public UserInfoDto findById(@PathVariable("id") Long id) {
+    public UserInfoDto findById(@PathVariable("id") Long id, @AuthenticationPrincipal UserDetailsImpl userDetails) {
+        if (!userDetails.isAdmin() && !userDetails.getId().equals(id)) {
+            throw new UserException(UserErrorResult.UNAUTHORIZED_READ_ATTEMPT);
+        }
         return userInfoService.findById(id);
     }
 
     @GetMapping
+    @PreAuthorize("hasAuthority('ROLE_ADMIN')")
     public List<UserInfoDto> findAll() {
         return userInfoService.findAll();
     }
 
     @PutMapping("{id}")
-    public UserInfoDto update(@PathVariable("id") Long id, @RequestBody UserInfoDto userDetailsDto) {
-        return userInfoService.update(id, userDetailsDto);
+    public UserInfoDto update(@PathVariable("id") Long id, @RequestBody UserInfoDto request, @AuthenticationPrincipal UserDetailsImpl userDetails) {
+        if (!userDetails.isAdmin() && !userDetails.getId().equals(id)) {
+            throw new UserException(UserErrorResult.UNAUTHORIZED_UPDATE_ATTEMPT);
+        }
+        return userInfoService.update(id, request);
     }
 
     @DeleteMapping("{id}")
-    public UserInfoDto delete(@PathVariable("id") Long id) {
+    public UserInfoDto delete(@PathVariable("id") Long id, @AuthenticationPrincipal UserDetailsImpl userDetails) {
+        if (!userDetails.isAdmin() && !userDetails.getId().equals(id)) {
+            throw new UserException(UserErrorResult.UNAUTHORIZED_DELETION_ATTEMPT);
+        }
         return userInfoService.delete(id);
     }
 }
